pyproject.toml
[project]
name = "agno-acroud"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.2.10",
    "ddgs>=9.9.0",
    "duckduckgo-search>=8.1.1",
    "fastapi[standard]>=0.121.1",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.43.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.2",
    "groq>=0.33.0",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.51.0",
    "tavily-python>=0.7.12",
    "yfinance>=0.2.66",
]
